all done

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dferreir <marvin@42lausanne.ch>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/25 14:09:59 by dferreir          #+#    #+#             */
-/*   Updated: 2022/10/28 16:11:17 by dferreir         ###   ########.fr       */
+/*   Updated: 2022/10/31 10:08:38 by dferreir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,11 +22,11 @@ int		ft_printchar(int c);
 
 int		ft_printhex(unsigned int nbr, char type);
 
-int		ft_printnbr(int	n);
+int		ft_printnbr(int n);
 
 int		ft_printunbr(unsigned int n);
 
-int		ft_printptr(unsigned long ptr);
+int		ft_printptr(unsigned long long ptr);
 
 int		ft_printstr(char *str);
